move mv bench to `test_helpers`, remove comparison
